Saga tests to match rework
